Move cityModule under appMainModule
