+++ app/app.less
@@ -19,8 +19,8 @@ body .main {
 @import "components/homePageModule/homePage";
 @import "components/appMainModule/aboutPageModule/aboutPage";
 @import "components/appMainModule/committeePageModule/styles";
+@import "components/appMainModule/cityPageModule/city";
 @import "components/appMainModule/faqPageModule/faqPage";
-@import "components/common/cityModule/city";
 
 //Angular UI Bootstrap Required Styling:
 .nav, .pagination, .carousel, .panel-title a { cursor: pointer; }
