+++ app/appRoutes.js
@@ -24,104 +24,6 @@ function appRoutes($stateProvider) {
     template: '<app-main></app-main>'
   };
 
-  var city = {
-    name: 'appMain.city',
-    abstract: true,
-    url: '^/city/:fips_id',
-    controller: require('./components/common/cityModule/cityController'),
-    template: require('./components/common/cityModule/index.html'),
-    resolve: {
-      city: function($stateParams, disclosureApi) {
-        return disclosureApi.locations.get({fips_id: $stateParams.fips_id});
-      }
-    },
-    data: {
-      moduleClasses: 'page',
-      pageClasses: 'city',
-      pageTitle: 'City',
-      pageDescription: 'Some description.'
-    }
-  };
-
-  var cityMoney = {
-    name: 'appMain.city.money',
-    url: '/money',
-    template: require('./components/common/cityModule/money/index.html'),
-    ncyBreadcrumb: {
-      label: '{{ city.location.name }}',
-      parent: 'appMain.city'
-    },
-    data: {
-      moduleClasses: 'page',
-      pageClasses: 'city',
-      pageTitle: 'City',
-      pageDescription: 'Some description.'
-    }
-  };
-
-  var cityElections = {
-    name: 'appMain.city.elections',
-    url: '/elections',
-    template: require('./components/common/cityModule/elections/index.html'),
-    controller: 'cityElectionsController',
-    ncyBreadcrumb: {
-      label: '{{ city.location.name }}',
-      parent: 'appMain.city'
-    },
-    resolve: {
-      ballot: function($q) {
-        return $q.resolve({
-          ballot_id: 'ballot1',
-          locality_id: 'locality2',
-          contests: [
-            {
-              contest_type: 'office',
-              name: 'Mayor'
-            },
-            {
-              contest_type: 'office',
-              name: 'City Auditor'
-            },
-            {
-              contest_type: 'office',
-              name: 'City Treasurer'
-            },
-            {
-              contest_type: 'office',
-              name: 'Distrit 1 City Council'
-            },
-            {
-              contest_type: 'office',
-              name: 'Distrit 3 City Council'
-            },
-            {
-              contest_type: 'office',
-              name: 'Distrit 5 City Council'
-            },
-            {
-              contest_type: 'referendum',
-              name: 'Measure AA'
-            },
-            {
-              contest_type: 'referendum',
-              name: 'Measure BB'
-            },
-            {
-              contest_type: 'referendum',
-              name: 'Measure CC'
-            }
-          ]
-        });
-      }
-    },
-    data: {
-      moduleClasses: 'page',
-      pageClasses: 'city',
-      pageTitle: 'City',
-      pageDescription: 'Some description.'
-    }
-  };
-
   var about = {
     name: 'appMain.about', // state name
     url: '^/about', // The ^ character makes this url override the parent url
@@ -189,9 +91,6 @@ function appRoutes($stateProvider) {
   $stateProvider.state(appMain);
   $stateProvider.state(about);
   $stateProvider.state(faq);
-  $stateProvider.state(city);
-  $stateProvider.state(cityMoney);
-  $stateProvider.state(cityElections);
   $stateProvider.state(examplePage1);
 
 }
