+++ app/components/appMainModule/appMain.js
@@ -1,10 +1,11 @@
 'use strict';
 
-var examplePage1Module = require('./examplePage1Module/examplePage1');
 var aboutPageModule = require('./aboutPageModule/aboutPage');
+var cityPageModule = require('./cityPageModule');
+var committeePageModule = require('./committeePageModule');
+var examplePage1Module = require('./examplePage1Module/examplePage1');
 var faqPageModule = require('./faqPageModule/faqPage');
 var measurePageModule = require('./measurePageModule');
-var committeePageModule = require('./committeePageModule');
 
 var appMainDirective = require('./appMainDirective');
 var AppMainController = require('./AppMainController');
@@ -14,6 +15,7 @@ module.exports = angular.module('appMainModule',
     'pageHeaderBreadcrumbsModule',
     'examplePage1Module',
     'aboutPageModule',
+    'cityPageModule',
     'committeePageModule',
     'faqPageModule',
     'measurePageModule'
