+++ app/components/appMainModule/cityPageModule/cityController.js
