+++ app/components/appMainModule/cityPageModule/cityElectionsController.js
