+++ app/components/components.js
@@ -2,10 +2,8 @@
 
 require('./homePageModule/homePage');
 require('./appMainModule/appMain');
-require('./common/cityModule');
 
 module.exports = angular.module('components', [
   'homePageModule',
-  'appMainModule',
-  'cityModule'
+  'appMainModule'
 ]);
